src/main/java/app/magicphoto/authservice/service/RoleService.java
package app.magicphoto.authservice.service;

import app.magicphoto.authservice.model.Role;
import app.magicphoto.authservice.repository.RoleRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class RoleService {

    private final RoleRepository roleRepo;

    @Autowired
    public RoleService(RoleRepository roleRepo) {
        this.roleRepo = roleRepo;
    }

    public Role findByName(String name) {
        return roleRepo.findByName(name).orElse(null);
    }
}
